chore: add defineRule
