--- docs/rules/issue-required-fields-content.md
@@ -1,7 +1,0 @@
-# `issue-required-fields-content`
-
-> Included in configs: ✅ recommended, 🔒 strict
-
-## See More
-
-- [Rule source](../../src/rules/issue-required-fields-content.ts)
