+++ src/action/comments/updateExistingCommentForReports.ts
@@ -1,6 +1,6 @@
 import type { EntityActor } from "../../actors/types.js";
 import type { CommentData, Entity } from "../../types/entities.js";
-import type { RuleReport } from "../../types/rules.js";
+import type { RuleReport } from "../../types/reports.js";
 
 import { markdownReporter } from "../../reporters/markdownReporter.js";
 import { createCommentBody } from "./createCommentBody.js";
