+++ src/reporters/cliReporter.test.ts
@@ -1,10 +1,11 @@
 import { describe, expect, test } from "vitest";
 
-import { RuleAbout, RuleReportData } from "../types/rules.js";
+import type { RuleReportData } from "../types/reports.js";
+import type { RuleAbout } from "../types/rules.js";
+
 import { cliReporter } from "./cliReporter.js";
 
 const fakeAbout = {
-	config: "recommended",
 	description: "Fake description.",
 	explanation: ["Fake explanation."],
 	name: "fake-rule",
