+++ src/rules/configs.ts
@@ -1,11 +1,11 @@
-import type { ConfigName } from "../types/configs.js";
+import type { ConfigName } from "../types/core.js";
 import type { Rule } from "../types/rules.js";
 
-import { rules } from "./all.js";
+import { allRules } from "./all.js";
 
 export const configs = {
-	recommended: rules.filter((rule) => rule.about.config === "recommended"),
-	strict: rules.filter((rule) =>
+	recommended: allRules.filter((rule) => rule.about.config === "recommended"),
+	strict: allRules.filter((rule) =>
 		["recommended", "strict"].includes(rule.about.config),
 	),
 } satisfies Record<ConfigName, Rule[]>;
