+++ src/rules/defineRule.ts
@@ -1,0 +1,5 @@
+import type { Rule, RuleAbout } from "../types/rules";
+
+export function defineRule<About extends RuleAbout>(rule: Rule<About>) {
+	return rule;
+}
