--- src/rules/issueRequiredFieldsContent.ts
@@ -1,13 +1,0 @@
-import type { Rule } from "../types/rules.js";
-
-export const issueRequiredFieldsContent = {
-	about: {
-		config: "recommended",
-		description: "TODO",
-		explanation: [],
-		name: "issue-required-fields-content",
-	},
-	issue(/* context, entity */) {
-		// TODO...
-	},
-} satisfies Rule;
