--- src/types/configs.ts
@@ -1,1 +1,0 @@
-export type ConfigName = "recommended" | "strict";
