+++ src/types/reports.ts
@@ -1,0 +1,39 @@
+import type { RuleAbout } from "./rules.js";
+
+/**
+ * Violation report generated by a rule for a GitHub entity.
+ */
+export interface RuleReport {
+	/**
+	 * Metadata describing the rule that generated the report.
+	 */
+	about: RuleAbout;
+
+	/**
+	 * Report data for this specific violation.
+	 */
+	data: RuleReportData;
+}
+
+/**
+ * Report data for a specific {@link RuleReport} violation.
+ */
+export interface RuleReportData {
+	/**
+	 * Single sentence description of the reported violation.
+	 * @example "The following image is missing alt text:"
+	 */
+	primary: string;
+
+	/**
+	 * Any additional sentences detailing the violation.
+	 * @example ["> [](img.jpg)"]
+	 */
+	secondary?: string[];
+
+	/**
+	 * Steps the user can take to resolve the violation.
+	 * @example "To resolve this report, add descriptive alt text to the image."
+	 */
+	suggestion: string[];
+}
